remove GpUnitClient enum, use REST client by default, GP-7220
