+++ .classpath
@@ -3,6 +3,7 @@
 	<classpathentry kind="src" path="src/main/java"/>
 	<classpathentry kind="src" path="src/test/java"/>
 	<classpathentry kind="src" path="soap/src/main/java"/>
+	<classpathentry kind="src" path="soap/src/test/java"/>
 	<classpathentry kind="lib" path="lib/gp-modules-updated.jar"/>
 	<classpathentry kind="lib" path="target/dependency/commons-codec-1.3.jar"/>
 	<classpathentry kind="lib" path="target/dependency/commons-compress-1.8.jar"/>
@@ -19,5 +20,6 @@
 	<classpathentry kind="lib" path="target/dependency/commons-logging-1.2.jar"/>
 	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER"/>
 	<classpathentry kind="con" path="org.eclipse.jdt.USER_LIBRARY/GENEPATTERN_CLIENT"/>
+	<classpathentry combineaccessrules="false" kind="src" path="/genepattern-server"/>
 	<classpathentry kind="output" path="classes"/>
 </classpath>
